+++ src/main/java/com/gbc/codingmates/controller/auth/AuthController.java
@@ -1,6 +1,6 @@
 package com.gbc.codingmates.controller.auth;
 
-import com.gbc.codingmates.config.oAuth.GoogleAuthService;
+import com.gbc.codingmates.api.oAuth.google.GoogleAuthService;
 import com.gbc.codingmates.dto.MemberDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
